libutils: Move BitStream into libutils
